Minor bug fixes for Refresh and Register actions

+++ api/actions/RefreshAction.php
@@ -3,10 +3,9 @@
 namespace yrc\api\actions;
 
 use yrc\api\actions\AuthenticationAction;
-use app\models\User\Token;
+use app\models\Token;
 use yrc\rest\Action as RestAction;
 
-use yii\web\UnauthorizedHttpException;
 use Yii;
 
 /**
@@ -16,7 +15,7 @@ use Yii;
 class RefreshAction extends RestAction
 {
     /**
-     * [POST] /api/v1/user/refresh
+     * [POST] /api/[...]/refresh
      * Refreshes the user's token
      * @return bool
      */
@@ -24,10 +23,22 @@ class RefreshAction extends RestAction
     {
         // Get the token
         $token = AuthenticationAction::getAccessTokenFromHeader();
-            
+        
+        $refreshToken = Yii::$app->request->post('refresh_token', false);
+
+        if ($refreshToken !== $token->refreshToken) {
+            return false;
+        }
+
         // If we can delete the token, send a newly generated token out
         if ($token->delete()) {
-            return Token::generate(Yii::$app->user->id);
+            $tokens = Token::generate(Yii::$app->user->id);
+            return [
+                'access_token'  => $tokens['accessToken'],
+                'refresh_token' => $tokens['refreshToken'],
+                'ikm'           => $tokens['ikm'],
+                'expires_at'    => $tokens['expiresAt']
+            ];
         }
 
         // Return false for any other reasons
